Change !say back to double-quotes
Discussed this a bit with Edwin, what I saw was likely just result of a
second Services bot also being in the same world and the name not being
exposed for the second line of text. This is a recent feature of VP that
allows chat to appear a bit cleaner in-universe.